+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
-
+- Added options on read_file.
+  [sgeulette]
 
 0.1 (2015-06-03)
 ----------------